+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version 1.1.0 (2018-07-30)
 
-* [new] OAuthRealm now provides security roles based on OAuth scopes.
+* [new] OAuthRealm now provides subject roles based on OAuth scopes.
+* [new] UserInfo data is now available as subject principals.
 
 # Version 1.0.0 (2018-05-04)